+++ cognite/neat/__init__.py
@@ -1,3 +1,4 @@
 from ._version import __version__
+from .utils.auth import get_cognite_client
 
-__all__ = ["__version__"]
+__all__ = ["__version__", "get_cognite_client"]