+++ cognite/neat/graph/loaders/_base.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
-from collections.abc import Iterable
+from collections.abc import Hashable, Iterable
 from pathlib import Path
-from typing import ClassVar, Generic, Literal, TypeVar, overload
+from typing import ClassVar, Generic, TypeVar
 
 from cognite.client import CogniteClient
 from cognite.client.data_classes.capabilities import Capability
@@ -10,7 +10,7 @@ from cognite.neat.graph import NeatGraphStore
 from cognite.neat.graph.issues.loader import FailedAuthorizationError
 from cognite.neat.issues import NeatIssue, NeatIssueList
 from cognite.neat.utils.auxiliary import class_html_doc
-from cognite.neat.utils.upload import UploadDiffsID, UploadResultIDs
+from cognite.neat.utils.upload import UploadResult, UploadResultList
 
 T_Output = TypeVar("T_Output")
 
@@ -43,45 +43,17 @@ class BaseLoader(ABC, Generic[T_Output]):
 class CDFLoader(BaseLoader[T_Output]):
     _UPLOAD_BATCH_SIZE: ClassVar[int] = 1000
 
-    @overload
-    def load_into_cdf_iterable(
-        self, client: CogniteClient, return_diffs: Literal[False] = False, dry_run: bool = False
-    ) -> Iterable[UploadResultIDs]: ...
+    def load_into_cdf(self, client: CogniteClient, dry_run: bool = False) -> UploadResultList:
+        return UploadResultList(self.load_into_cdf_iterable(client, dry_run))
 
-    @overload
-    def load_into_cdf_iterable(
-        self, client: CogniteClient, return_diffs: Literal[True], dry_run: bool = False
-    ) -> Iterable[UploadDiffsID]: ...
-
-    def load_into_cdf_iterable(
-        self, client: CogniteClient, return_diffs: bool = False, dry_run: bool = False
-    ) -> Iterable[UploadResultIDs] | Iterable[UploadDiffsID]:
-        yield from self._load_into_cdf_iterable(client, return_diffs, dry_run)
-
-    @overload
-    def load_into_cdf(
-        self, client: CogniteClient, return_diffs: Literal[False] = False, dry_run: bool = False
-    ) -> list[UploadResultIDs]: ...
-
-    @overload
-    def load_into_cdf(
-        self, client: CogniteClient, return_diffs: Literal[True], dry_run: bool = False
-    ) -> list[UploadDiffsID]: ...
-
-    def load_into_cdf(
-        self, client: CogniteClient, return_diffs: bool = False, dry_run: bool = False
-    ) -> list[UploadResultIDs] | list[UploadDiffsID]:
-        return list(self._load_into_cdf_iterable(client, return_diffs, dry_run))  # type: ignore[return-value]
-
-    def _load_into_cdf_iterable(
-        self, client: CogniteClient, return_diffs: bool = False, dry_run: bool = False
-    ) -> Iterable[UploadResultIDs] | Iterable[UploadDiffsID]:
+    def load_into_cdf_iterable(self, client: CogniteClient, dry_run: bool = False) -> Iterable[UploadResult]:
         missing_capabilities = client.iam.verify_capabilities(self._get_required_capabilities())
-        result_cls = UploadDiffsID if return_diffs else UploadResultIDs
         if missing_capabilities:
-            result = result_cls(name=type(self).__name__)
-            result.issues.append(FailedAuthorizationError(action="Upload to CDF", reason=str(missing_capabilities)))
-            yield result
+            upload_result = UploadResult[Hashable](name=type(self).__name__)
+            upload_result.issues.append(
+                FailedAuthorizationError(action="Upload to CDF", reason=str(missing_capabilities))
+            )
+            yield upload_result
             return
 
         issues = NeatIssueList[NeatIssue]()
@@ -93,10 +65,11 @@ class CDFLoader(BaseLoader[T_Output]):
                 items.append(result)
 
             if len(items) >= self._UPLOAD_BATCH_SIZE:
-                yield self._upload_to_cdf(client, items, return_diffs, dry_run, issues)
-                items.clear()
+                yield self._upload_to_cdf(client, items, dry_run, issues)
+                issues = NeatIssueList[NeatIssue]()
+                items = []
         if items:
-            yield self._upload_to_cdf(client, items, return_diffs, dry_run, issues)
+            yield self._upload_to_cdf(client, items, dry_run, issues)
 
     @abstractmethod
     def _get_required_capabilities(self) -> list[Capability]:
@@ -107,8 +80,7 @@ class CDFLoader(BaseLoader[T_Output]):
         self,
         client: CogniteClient,
         items: list[T_Output],
-        return_diffs: bool,
         dry_run: bool,
         read_issues: NeatIssueList,
-    ) -> UploadResultIDs | UploadDiffsID:
+    ) -> UploadResult:
         raise NotImplementedError